app/requirements.txt
# fastapi
# uvicorn[standard]
# sqlalchemy[asyncio]
# asyncpg
# alembic


fastapi==0.115.14
uvicorn[standard]==0.35.0
python-multipart==0.0.16
sqlalchemy==2.0.41
alembic==1.16.2
asyncpg==0.30.0
aiosqlite==0.19.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
jinja2==3.1.4
aiosmtplib==3.0.2
email-validator==2.1.1
pillow==11.2.1
aiofiles==24.1.0
python-magic==0.4.27
pydantic==2.11.7
pydantic-settings==2.3.4
python-dotenv==1.1.1
qrcode[pil]==8.0
pandas==2.1.4
plotly==5.18.0
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.27.2
